% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_pobierz_fb_likes.R
\name{wybory_pobierz_fb_likes}
\alias{wybory_pobierz_fb_likes}
\title{Pobieranie liczby polubien z fanpage'y kandydatow}
\usage{
\code{wybory_pobierz_fb_likes(sciezka_zapis=getwd())}
}
\arguments{
\item{sciezka_zapis}{sciezka do miejsca na dysku, gdzie ma zostac zapisany plik z pobranymi polubieniami}
}
\value{
Zwraca invisible NULL.
}
\description{
Funkcja \code{wybory_pobierz_fb_likes} pobiera aktualna liczbe polubien z fanpage'y kandydatow.
}
\details{
Funkcja wyciaga z kodu zrodlowego fanpage'ow kandydatow liczbe polubien.

Pliki zostana zapisane we wskazanym miejscu na dysku w folderze Facebook/likes/, a
nazwa pliku bedzie w formacie "facebook_likes-(czas systemowy).txt".
}
\examples{
wybory_pobierz_fb_likes()
}
\author{
Piotr Smuda
}

