% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_pobierz_fb_posty.R
\name{wybory_pobierz_fb_posty}
\alias{wybory_pobierz_fb_posty}
\title{Pobieranie postow z fanpage'y kandydatow}
\usage{
\code{wybory_pobierz_fb_posty(sciezka_zapis=getwd(),sciezka_dostep_api)}
}
\arguments{
\item{sciezka_zapis}{sciezka do miejsca na dysku, gdzie ma zostac zapisany plik z pobranymi postami}

\item{sciezka_dostep_api}{sciezka do pliku z pelna autoryzacja do aplikacji facebook}
}
\value{
Zwraca invisible NULL oraz komunikat z informacja o wynikach sesji pobierania.
}
\description{
Funkcja \code{wybory_pobierz_fb_posty} pobiera posty z fanpage'y kandydatow z dnia dzisiejszego.
}
\details{
Funkcja dzieki dostepowi do aplikacji na facebooku wyciaga wszystkie dostepne informacje o postach
z fanpage'ow kandydatow.

Pliki zostana zapisane we wskazanym miejscu na dysku w folderze Facebook/posty/(imie i nazwisko kandydata)/, a
nazwa pliku bedzie w formacie "posty-(data systemowa).txt".

KROTKA INSTRUKCJA ZAPISU PLIKU Z AUTORYZACJA DO APLIKACJI NA FACEBOOKU

1. Na poczatku trzeba stworzyc aplikacje API na facebooku.

2. Potem tworzy sie zmienna srodowiskowa np. fb_oauth
   przy pomocy funkcji fbOAuth(app_id,app_secret,extended_permissions=TRUE).

3. Nastepnie zapisuje sie ja do pliku przy pomocy funkcji save(fb_oauth,file).
}
\author{
Piotr Smuda
}

