% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_pobierz_GoogleNews.R
\name{wybory_pobierz_GoogleNews}
\alias{wybory_pobierz_GoogleNews}
\title{Pobieranie informacji z portalu GoogleNews}
\usage{
wybory_pobierz_GoogleNews(sciezka_zapis=getwd())
}
\arguments{
\item{sciezka_zapis}{sciezka do folderu w ktorym maja zostac zapisane artykuly}
}
\value{
Zwraca invisible NULL oraz komunikat z informacja o wynikach sesji pobierania.
}
\description{
Funkcja \code{wybory_pobierz_GoogleNews} zlicza wystapienia nazwisk kandydatow
we wszystkich artykulach na stronie GoogleNews : http://news.google.pl/.
}
\details{
Artykuly na tym portalu naleza do roznych stron, a kazda z nich ma swoja
organizacje tekstu, dlatego ponizszy kod wyciaga caly kod HTML strony
z artykulem i zlicza ile razy wystepuja w nim nazwiska kandydatow.

Pliki zostana zapisane we wskazanym miejscu na dysku w folderze GoogleNews/, a
nazwa pliku bedzie w formacie "GoogleNews(czas systemowy).txt".
}
\examples{
wybory_pobierz_GoogleNews()
}
\author{
Karolina Wyszynska
}

