% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_pobierz_Interia.R
\name{wybory_pobierz_Interia}
\alias{wybory_pobierz_Interia}
\title{Pobieranie informacji z portalu Interia}
\usage{
wybory_pobierz_Interia(sciezka_zapis=getwd())
}
\arguments{
\item{sciezka_zapis}{sciezka do folderu w ktorym maja zostac zapisne artykuly}
}
\value{
Zwraca invisible NULL oraz komunikat z informacja o wynikach sesji pobierania.
}
\description{
Funkcja \code{wybory_pobierz_Interia} pobiera artykuly z portalu Interia,
a dokladniej ze strony dedykowanej dla wyborow:
"http://fakty.interia.pl/raport-wybory-prezydenckie-2015".
}
\details{
Korzystamy ze strony informacyjnej Interii z wyselekcjonowanymi
informacjami na temat wyborow. Artykuly, ktore sa tam podlinkowane maja podobna
strukture, dlatego wyciagamy z nich caly tekst artykulu i zapisujemy go
do pliku.

Pliki zostana zapisane we wskazanym miejscu na dysku w folderze Interia/(czas systemowy)/, a
nazwa pliku bedzie w formacie liczb calkowitych od 1 do 10 z rozszerzeniem *.txt.
}
\examples{
wybory_pobierz_Interia()
}
\author{
Karolina Wyszynska
}

