% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_pobierz_tw.R
\name{wybory_pobierz_tw}
\alias{wybory_pobierz_tw}
\title{Pobieranie tweetow o kandydatach}
\usage{
\code{wybory_pobierz_tw(sciezka_zapis=getwd(),sciezka_dostep_api,czas)}
}
\arguments{
\item{sciezka_zapis}{sciezka do miejsca na dysku, gdzie ma zostac zapisany plik z pobranymi tweetami}

\item{sciezka_dostep_api}{sciezka do pliku z pelna autoryzacja do aplikacji twitter}

\item{czas}{ilosc godzin dzialania funkcji}
}
\value{
Zwraca invisible NULL oraz komunikat z informacja o wynikach sesji pobierania.
}
\description{
Funkcja \code{wybory_pobierz_tw} pobiera tweety ze slowami ze slownik_tw w czasie rzeczywistym.
}
\details{
Funkcja dzieki dostepowi do aplikacji na twitterze pobiera w czasie rzeczywistym pojawiajace sie tweety.

Pliki zostana zapisane we wskazanym miejscu na dysku w folderze Twitter/tweety/, a
nazwa pliku bedzie w formacie "tweety-(data systemowa).json".

KROTKA INSTRUKCJA ZAPISU PLIKU Z AUTORYZACJA DO APLIKACJI NA TWITTERZE

1. Na poczatku trzeba stworzyc aplikacje API na Twitterze.

2. Potem tworzy sie jakas zmienna srodowiskowa np. paczka
  przy pomocy funkcji kodu:

requestURL<-"https://api.twitter.com/oauth/request_token"

accessURL<-"https://api.twitter.com/oauth/access_token"

authURL<-"https://api.twitter.com/oauth/authorize"

consumerKey<-"XXXXXXXXXXXXXXXXXXXXXXXXXX"

consumerSecret<-"XXXXXXXXXXXXXXXXXXXXXXXXXX"

oauthKey<-"XXXXXXXXXXXXXXXXXXXXXXXXXX"

oauthSecret<-"XXXXXXXXXXXXXXXXXXXXXXXXXX"

tw_oauth<-OAuthFactory$new(consumerKey=consumerKey,consumerSecret=consumerSecret,
     oauthKey=oauthKey,oauthSecret=oauthSecret,requestURL=requestURL,
     accessURL=accessURL,authURL=authURL)

3. Autoryzujemy manualnie poprzez:
tw_oauth$handshake(cainfo=system.file("CurlSSL","cacert.pem",package="RCurl")).

4. Nastepnie zapisuje sie zmienna do pliku przy pomocy funkcji save(tw_oauth,file).
}
\author{
Piotr Smuda
}

