% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_pobierz_WP.R
\name{wybory_pobierz_WP}
\alias{wybory_pobierz_WP}
\title{Pobieranie informacji z portalu WP}
\usage{
wybory_pobierz_WP(sciezka_zapis=getwd())
}
\arguments{
\item{sciezka_zapis}{sciezka do folderu w ktorym maja zostac zapisne artykuly}
}
\value{
Zwraca invisible NULL oraz komunikat z informacja o wynikach sesji pobierania.
}
\description{
Funkcja \code{wybory_pobierz_WP} pobiera artykuly z portalu WP,
a dokladniej ze strony dedykowanej dla wyborow:
"http://wiadomosci.wp.pl/kat,140394,title,Wybory-prezydenckie-w-2015-r,raport.html".
}
\details{
Korzystamy ze strony informacyjnej Interii z wyselekcjonowanymi
informacjami na temat wyborow. Artykuly, ktore sa tam podlinkowane maja podobna
strukture, dlatego wyciagamy z nich caly tekst artykulu i zapisujemy go
do pliku.

Pliki zostana zapisane we wskazanym miejscu na dysku w folderze WirtualnaPolska/(czas systemowy)/, a
nazwa pliku bedzie w formacie liczb calkowitych od 1 do 18 z rozszerzeniem *.txt.
}
\examples{
wybory_pobierz_WP()
}
\author{
Karolina Wyszynska
}

