% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_stworz_baze.R
\name{wybory_stworz_baze}
\alias{wybory_stworz_baze}
\title{Tworzenie bazy ze wskaznikami}
\usage{
\code{wybory_stworz_baze(sciezka_do_bazy)}
}
\arguments{
\item{sciezka_do_bazy}{sciezka do pliku, w ktorym ma zostac zapisana baza}
}
\value{
Zwraca invisible NULL.
}
\description{
Funkcja \code{wybory_stworz_baze} tworzy szkielet bazy relacyjnej ze wskaznikami dla kandydatow.
}
\details{
Utworzona baza ze wskaznikami zawiera tabelki:

1. Kandydaci o kolumnach IDKandydata (klucz glowny, klucz obcy), ImieINazwisko.

2. Dzien o kolumnach Wzorzec (klucz glowny), IDKandydata (klucz glowny, klucz obcy),
IloscTweetow, WydzwiekTweetow, IloscPolubienFB, SzybkoscPolubienFB, IloscPostowFB,
SredniaLikePostFB, SredniaKomentarzPostFB, SredniaSharePostFB, IloscGoogle, SredniGoogle,
TytulyInteria, TytulyOnet, TytulyWP, TytulyPortale, ArtykulyInteria, ArtykulyOnet,
ArtykulyWP, ArtykulyPortale.

3. Tydzien o kolumnach Wzorzec (klucz glowny), IDKandydata (klucz glowny, klucz obcy),
IloscTweetow, WydzwiekTweetow, IloscPolubienFB, SzybkoscPolubienFB, IloscPostowFB,
SredniaLikePostFB, SredniaKomentarzPostFB, SredniaSharePostFB, IloscGoogle, SredniGoogle,
TytulyInteria, TytulyOnet, TytulyWP, TytulyPortale, ArtykulyInteria, ArtykulyOnet,
ArtykulyWP, ArtykulyPortale.

4. Miesiac o kolumnach Wzorzec (klucz glowny), IDKandydata (klucz glowny, klucz obcy),
IloscTweetow, WydzwiekTweetow, IloscPolubienFB, SzybkoscPolubienFB, IloscPostowFB,
SredniaLikePostFB, SredniaKomentarzPostFB, SredniaSharePostFB, IloscGoogle, SredniGoogle,
TytulyInteria, TytulyOnet, TytulyWP, TytulyPortale, ArtykulyInteria, ArtykulyOnet,
ArtykulyWP, ArtykulyPortale.
}
\author{
Piotr Smuda
}

