% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_wsk_artykuly.R
\name{wybory_wsk_artykuly}
\alias{wybory_wsk_artykuly}
\title{Wskaznik widocznosci  kandydata - portale internetowe}
\usage{
wybory_wsk_artykuly(okres, wzor, serwis, sciezka_do_folderu=getwd())
}
\arguments{
\item{okres}{okres, za ktory jest liczony wskaznik; mozliwosci: "dzien", "tydzien", "miesiac"}

\item{wzor}{wzorzec do odczytu plikow odpowiadajacy danemu okresowi; mozliwosci: "2015-XX-XX", "2015-(X|XX)", "2015-XX"}

\item{serwis}{nazwa serwisu, dla ktorego obliczany ma byc wskaznik: "Interia", "Onet", "WP"}

\item{sciezka_do_folderu}{sciezka do folderu, ktory zawiera podfolder z wynikami dzialania funkcji wybory_pobierz_WP(), wybory_pobierz_Interia() lub wybory_pobierz_Onet()}
}
\value{
Dla danego okresu zwraca wyliczony wskaznik  w postaci wektora nazwanego nazwiskami poszczegolnych kandydatow.
}
\description{
Funkcja \code{wybory_wsk_artykuly} liczy wskaznik widocznosci kandydatow na portalach informacyjnych.
}
\details{
Dla kazdego artykulu oraz kazdego kandydata przeprowadzamy analize tekstowa
zdan, w ktorych wystepuje nazwisko kandydata. (Gdy zdan nie odnaleziono
wskaznik wynosi 0). Sumujemy oceny slow razy ilosc ich wystepowania i dzielimy
wynik przez ilosc ocenionych slow. Dostajemy wskaznik z przedzialu [1,5]
(Gdy brak slow do oceny - wstawiamy wskaznik neutralny = 3)
Nastepnie przemnazamy to przez range artykulu (te wyzej na stronie sa bardziej
widoczne) i sumujemy po wszystkich artykulach.
}
\examples{
wybory_wsk_artykuly("dzien", "2015-03-23", "Interia")
wybory_wsk_artykuly("dzien", "2015-03-23", "WirtualnaPolska")
wybory_wsk_artykuly("miesiac", "2015-04", "Onet")
}
\author{
Karolina Wyszynska
}

