% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_wsk_fb_liczba_like_post.R
\name{wybory_wsk_fb_liczba_like_post}
\alias{wybory_wsk_fb_liczba_like_post}
\title{Wskaznik sredniej liczby polubien na posta na fanpage'u kandydata}
\usage{
\code{wybory_wsk_fb_liczba_like_post(okres,wzor,sciezka_odczyt=getwd())}
}
\arguments{
\item{okres}{okres, za ktory jest liczony wskaznik; mozliwosci: "dzien", "tydzien", "miesiac"}

\item{wzor}{wzor do odczytu plikow odpowiadajacy danemu okresowi; mozliwosci: "2015-XX-XX", "2015-(X|XX)", "2015-XX"}

\item{sciezka_odczyt}{sciezka do folderu bazowego z wszystkimi plikami}
}
\value{
Dla danego okresu zwraca wyliczony wskaznik  w postaci wektora nazwanego nazwiskami poszczegolnych kandydatow.
}
\description{
Funkcja \code{wybory_wsk_fb_liczba_like_post} zlicza srednia liczbe polubien na posta na fanpage'u kandydata.
}
\details{
Sposob obliczania wskaznika "sredniej ilosci" polubien postow na fanpage'u:

Dla danego okresu zlicza srednia arytmetyczna liczby polubien postow
opublikowanych przez kandydata.
}
\examples{
wybory_wsk_fb_liczba_like_post("dzien","2015-04-14")
wybory_wsk_fb_liczba_like_post("tydzien","2015-14")
wybory_wsk_fb_liczba_like_post("miesiac","2015-04")
}
\author{
Piotr Smuda
}

