% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_wsk_fb_liczba_postow.R
\name{wybory_wsk_fb_liczba_postow}
\alias{wybory_wsk_fb_liczba_postow}
\title{Wskaznik ilosci postow na fanpage'u kandydata}
\usage{
\code{wybory_wsk_fb_liczba_postow(okres,wzor,sciezka_odczyt=getwd())}
}
\arguments{
\item{okres}{okres, za ktory jest liczony wskaznik; mozliwosci: "dzien", "tydzien", "miesiac"}

\item{wzor}{wzor do odczytu plikow odpowiadajacy danemu okresowi; mozliwosci: "2015-XX-XX", "2015-(X|XX)", "2015-XX"}

\item{sciezka_odczyt}{sciezka do folderu bazowego z wszystkimi plikami}
}
\value{
Dla danego okresu zwraca wyliczony wskaznik  w postaci wektora nazwanego nazwiskami poszczegolnych kandydatow.
}
\description{
Funkcja \code{wybory_wsk_fb_liczba_postow} zlicza ilosci postow na fanpage'u kandydata.
}
\details{
Sposob obliczania wskaznika liczby postow na fanpage'u:

Dla danego okresu zlicza liczbe postow opublikowanych przez kandydata.
}
\examples{
wybory_wsk_fb_liczba_postow("dzien","2015-04-14")
wybory_wsk_fb_liczba_postow("tydzien","2015-14")
wybory_wsk_fb_liczba_postow("miesiac","2015-04")
}
\author{
Piotr Smuda
}

