% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_wsk_GoogleNews.R
\name{wybory_wsk_GoogleNews}
\alias{wybory_wsk_GoogleNews}
\title{Wskazniki widocznosci kandydata w GoogleNews}
\usage{
wybory_wsk_GoogleNews(okres, wzor, typ="zwykly", sciezka_do_folderu=getwd())
}
\arguments{
\item{okres}{okres, za ktory jest liczony wskaznik; mozliwosci: "dzien", "tydzien", "miesiac"}

\item{wzor}{wzorzec do odczytu plikow odpowiadajacy danemu okresowi; mozliwosci: "2015-XX-XX", "2015-(X|XX)", "2015-XX"}

\item{typ}{okresla typ wskaznika: "zwykly", "sredni"}

\item{sciezka_do_folderu}{sciezka do folderu, ktory zawiera podfolder GoogleNews (wynik dzialania funkcji wybory_pobierz_GoogleNews())}
}
\value{
Dla danego okresu zwraca wyliczony wskaznik  w postaci wektora nazwanego nazwiskami poszczegolnych kandydatow.
}
\description{
Funkcja \code{wybory_wsk_google} zlicza wskazniki
widocznosci kandydatow na portalu informacyjnym GoogleNews - wiecej patrz
funkcja wybory_pobierz_GoogleNews().
}
\details{
Sposob obliczania wskaznika "sredniego" wyswietlenia na Google:

Zwraca sume wyswietlen danego nazwiska na portalu dla danego okresu czasu.
}
\examples{
wybory_wsk_GoogleNews("dzien", "2015-03-23")
wybory_wsk_GoogleNews("tydzien", "2015-12")
wybory_wsk_GoogleNews("miesiac", "2015-03")
wybory_wsk_GoogleNews("dzien", "2015-03-23", "sredni")
wybory_wsk_GoogleNews("tydzien", "2015-12", "sredni")
wybory_wsk_GoogleNews("miesiac", "2015-04", "sredni")
}
\author{
Karolina Wyszynska
}

