% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_wsk_tw_analiza_wydzwieku.R
\name{wybory_wsk_tw_analiza_wydzwieku}
\alias{wybory_wsk_tw_analiza_wydzwieku}
\title{Wskaznik wydzwieku tweetow z kandydatami}
\usage{
\code{wybory_wsk_tw_analiza_wydzwieku(okres,wzor,sciezka_odczyt=getwd())}
}
\arguments{
\item{okres}{okres, za ktory jest liczony wskaznik; mozliwosci: "dzien", "tydzien", "miesiac"}

\item{wzor}{wzor do odczytu plikow odpowiadajacy danemu okresowi; mozliwosci: "2015-XX-XX", "2015-(X|XX)", "2015-XX"}

\item{sciezka_odczyt}{sciezka do folderu bazowego z wszystkimi plikami}
}
\value{
Dla danego okresu zwraca wyliczony wskaznik  w postaci wektora nazwanego nazwiskami poszczegolnych kandydatow.
}
\description{
Funkcja \code{wybory_wsk_tw_analiza_wydzwieku} zlicza wydzwiek tweetow z kandydatami.
}
\details{
Sposob obliczania wskaznika wydzwieku tweetow o kandydatach na Twitterze:

Opisuje wydzwiekk tweetow zebranych w danym okresie, w ktorych znaleziono nazwisko
danego kandydata. Dla konkretnego kandydata wyszukujemy wszystkie tweety z jego
nazwiskiem i oceniamy wartosc poszczegolnych slow wystepujacych w tych tweetach.
Sumujemy te wartosci. By uzyskac wartosc z przedzialu [1, 5] dzielimy to przez
ilosc ocenionych wyrazow. Jesli nie wykryto, zadnego zabarwionego emocjonalnie
slowa tweet uznajemy za neutralny i otrzymuje wartosc 3. Gdy nie znalezlismy
zadnych tweetow dotyczacych danego kandydata przyjmujemy wskaznik rowny 0.
}
\examples{
wybory_wsk_tw_analiza_wydzwieku("dzien","2015-04-14")
wybory_wsk_tw_analiza_wydzwieku("tydzien","2015-14")
wybory_wsk_tw_analiza_wydzwieku("miesiac","2015-04")
}
\author{
Karolina Wyszynska
}

