% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_wsk_tw_liczba_wystapien.R
\name{wybory_wsk_tw_liczba_wystapien}
\alias{wybory_wsk_tw_liczba_wystapien}
\title{Wskaznik liczby wystapien kandydata w tweetach}
\usage{
\code{wybory_wsk_tw_liczba_wystapien(okres,wzor,sciezka_odczyt=getwd())}
}
\arguments{
\item{okres}{okres, za ktory jest liczony wskaznik; mozliwosci: "dzien", "tydzien", "miesiac"}

\item{wzor}{wzorzec do odczytu plikow odpowiadajacy danemu okresowi; mozliwosci: "2015-XX-XX", "2015-(X|XX)", "2015-XX"}

\item{sciezka_odczyt}{sciezka do folderu bazowego z wszystkimi plikami}
}
\value{
Dla danego okresu zwraca wyliczony wskaznik  w postaci wektora nazwanego nazwiskami poszczegolnych kandydatow.
}
\description{
Funkcja \code{wybory_wsk_tw_liczba_wystapien} zlicza liczbe wystapien kandydata w tweetach.
}
\details{
Sposob obliczania wskaznika liczby wystapien kandydata w tweetach na Twitterze:

Dla danego okresu dla kazdego tweeta zlicza, jaki kandydat w nim sie pojawil,
skleja wszystko do ramki danych, a nastepnie sumuje po kolumnach (kandydaci)
liczbe wystapien.
}
\examples{
wybory_wsk_tw_liczba_wystapien("dzien","2015-04-14")
wybory_wsk_tw_liczba_wystapien("tydzien","2015-14")
wybory_wsk_tw_liczba_wystapien("miesiac","2015-04")
}
\author{
Piotr Smuda
}

