% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_wsk_tytuly.R
\name{wybory_wsk_tytuly}
\alias{wybory_wsk_tytuly}
\title{Wskaznik widocznosci kandydata - portale internetowe - tytuly}
\usage{
wybory_wsk_tytuly(okres, wzor, serwis, sciezka_do_folderu=getwd())
}
\arguments{
\item{okres}{okres, za ktory jest liczony wskaznik; mozliwosci: "dzien", "tydzien", "miesiac"}

\item{wzor}{wzorzec do odczytu plikow odpowiadajacy danemu okresowi; mozliwosci: "2015-XX-XX", "2015-(X|XX)", "2015-XX"}

\item{serwis}{nazwa serwisu, dla ktorego obliczany ma byc wskaznik: "Interia", "Onet", "WP"}

\item{sciezka_do_folderu}{sciezka do folderu, ktory zawiera podfolder z wynikami dzialania funkcji wybory_pobierz_WP(), wybory_pobierz_Interia() lub wybory_pobierz_Onet()}
}
\value{
Dla danego okresu zwraca wyliczony wskaznik  w postaci wektora nazwanego nazwiskami poszczegolnych kandydatow.
}
\description{
Funkcja \code{wybory_wsk_tytuly} zlicza widocznosc kandydatow w tytulach na portalach informacyjnych.
}
\details{
Po podaniu portalu okresu, dla ktorego chcemy obliczyc wskaznik, oraz wzorca
okresu funkcja zczytuje pliki z danymi dotyczacymi tego okresu i analizuje
poszczegolne tytulu artykulow.

Sposob obliczania wskaznika:

Zliczana jest ilosc wystapien w tytulach artykulow i mnozona przez wage artykulu.
}
\examples{
wybory_wsk_tytuly("dzien", "2015-03-23", "Interia")
wybory_wsk_tytuly("dzien", "2015-03-23", "WirtualnaPolska")
wybory_wsk_tytuly("miesiac", "2015-04", "Onet")
}
\author{
Karolina Wyszynska
}

