% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_wskazniki_do_bazy.R
\name{wybory_wskazniki_do_bazy}
\alias{wybory_wskazniki_do_bazy}
\title{Uzupelnianie bazy ze wskaznikami}
\usage{
\code{wybory_wskazniki_do_bazy(wzor,okres,sciezka_do_bazy,sciezka_odczyt=getwd())}
}
\arguments{
\item{wzor}{wzor do odczytu plikow odpowiadajacy danemu okresowi; mozliwosci: "2015-XX-XX", "2015-(X|XX)", "2015-XX"}

\item{okres}{okres, za ktory jest liczony wskaznik; mozliwosci: "dzien", "tydzien", "miesiac"}

\item{sciezka_do_bazy}{sciezka do pliku z baza, ktora ma zostac uzupelniona}

\item{sciezka_odczyt}{sciezka do folderu bazowego z wszystkimi plikami zawierajacymi dane}
}
\value{
Zwraca invisible NULL.
}
\description{
Funkcja \code{wybory_wskazniki_do_bazy} uzupelnia baze relacyjna wyliczonymi wskaznikami dla wskazanego okresu.
}
\details{
Dla podanego okresu funkcja ta oblicza wskazniki i wstawia je do bazy.
}
\author{
Piotr Smuda, Karolina Wyszynska
}

