% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wybory_wstaw_do_bazy.R
\name{wybory_wstaw_do_bazy}
\alias{wybory_wstaw_do_bazy}
\title{Uzupelnianie bazy ze wskaznikami}
\usage{
\code{wybory_wstaw_do_bazy(okres,sciezka_do_bazy,sciezka_odczyt=getwd())}
}
\arguments{
\item{okres}{okres, za ktory jest liczony wskaznik; mozliwosci: "dzien", "tydzien", "miesiac"}

\item{sciezka_do_bazy}{sciezka do pliku, w ktorym jest baza, ktora zostanie uzupelniona}

\item{sciezka_odczyt}{sciezka do folderu bazowego z wszystkimi plikami}
}
\value{
Zwraca invisible NULL.
}
\description{
Funkcja \code{wybory_wstaw_do_bazy} uzupelnia baze relacyjna wyliczonymi wskaznikami dla okresu bez analizy.
}
\details{
Dzieki funkcjom pomocnicznym sprawdzany jest okres, w ktorym byla
przeprowadzona ostatnia analiza, a nastepnie dla brakujacych okresow zostaja
obliczone i wstawione wskazniki do bazy ze wskaznikami.
}
\author{
Piotr Smuda, Karolina Wyszynska
}

